% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/external.bcdm_to_dnastringset.R
\name{bcdm_to_dnastringset}
\alias{bcdm_to_dnastringset}
\title{Convert the BOLD parquet search into a DNAStringSet object}
\usage{
bcdm_to_dnastringset(bold.search.res, marker = NULL, cols_for_seq_names)
}
\arguments{
\item{bold.search.res}{A \code{tbl_sql} object containing BOLD search results}

\item{marker}{Character vector specifying the genetic marker}

\item{cols_for_seq_names}{Character vector of field names to include in the header (separated by "|")}
}
\value{
A \code{DNAStringSet} object
}
\description{
Converts the sequence data from the search result into a DNAStringSet object for downstream multiple sequence alignment with customized headers.
}
\details{
This function transforms the search results from \code{bold_parquet_search} into a \code{DNAStringSet} object suitable for downstream data analyses. The \code{cols_for_seq_names} argument lets users create custom headers using the BCDM column names.
}
\examples{
\dontrun{


# Search the BOLD data
bold_search <- bold_parquet_search(
input.parquet=parquet_file,
taxonomy = "Coleoptera",
geography = "Canada",
basecount = c(500, 660)
)

# Get the DNAStringset object

bold.dnastringset<-bcdm_to_dnastringset(bold_search,
marker="COI-5P",
cols_for_seq_names = c("processid","family")
)

}

}
