% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/external.bcdm_to_dwc.R
\name{bcdm_to_dwc}
\alias{bcdm_to_dwc}
\title{Convert the BOLD parquet search in to a Darwin Core (DwC) format data model}
\usage{
bcdm_to_dwc(bold.search.res)
}
\arguments{
\item{bold.search.res}{A \code{tbl_sql} object containing BOLD search results}
}
\value{
A data frame with columns mapped to Darwin Core equivalent fields
}
\description{
Converts \code{bold_parquet_search} results from BCDM format to Darwin Core Standard format.
}
\details{
This function maps BCDM (Barcode Core Data Model) fields to their Darwin Core equivalents.

\emph{Important Note}: All fields should be available in the \code{bold_parquet_search} \code{tbl_sql} object otherwise, the function will throw an error
}
\examples{
\dontrun{

# Search the BOLD data
bold_search <- bold_parquet_search(
  input.parquet = parquet_file,
  taxonomy = c("Cerambycidae", "Meloidae"),
  geography = "Australia"
)

# Get the DNAStringset object

bold.dwc <- bcdm_to_dwc(bold_search)
}
}
