% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/external.bcdm_to_occmatrix.R
\name{bcdm_to_occmatrix}
\alias{bcdm_to_occmatrix}
\title{Convert the BOLD parquet search into an occurrence matrix}
\usage{
bcdm_to_occmatrix(
  bold.search.res,
  kingdom = "Animalia",
  taxon.rank,
  taxon.name = NULL,
  site.cat = NULL,
  pre.abs = FALSE
)
}
\arguments{
\item{bold.search.res}{A \code{tbl_sql} object containing BOLD search results.}

\item{kingdom}{Character value specifying the kingdom (default: Animalia).}

\item{taxon.rank}{Taxonomic rank to aggregate by (kingdom, phylum, class, order, family, genus, species or bin_uri).}

\item{taxon.name}{Optional vector of specific taxon names to include (e.g., for \code{taxon.rank} = 'class', name can be 'Insecta').}

\item{site.cat}{Optional categorical variable to group occurrence data by (e.g., region; when \code{site.cat} = NULL,\code{coord} used as default)}

\item{pre.abs}{Logical indicating whether to convert counts to presence/absence (1/0) data (default: FALSE).}
}
\value{
A data frame with occurrence data (taxon names as columns, site categories or coordinates as rows).
}
\description{
Extracts occurrence data (specimen counts by taxon and location) from BOLD search results.
}
\details{
This function transforms the search results from \code{bold_parquet_search} into the occurrence data matrices commonly used in biodiversity and ecological analyses by packages like \code{vegan} and \code{betapart}. Occurrences differ based on the kingdom. For \code{Animalia}, only records with BINs are included. For other kingdoms, all records with a sequence are counted (i.e., records without sequences are removed before calculations). Records can be aggregated at different taxonomic ranks (from kingdom to BINs) for a single or multiple taxa, with optional filtering by specific taxon names. \code{site.cat} can be any of the \code{geography} fields. The function can convert count data to presence/absence (1/0) format.
\emph{Important Note}: The \code{bcdm_to_occmatrix} function requires \code{taxonomy} (including \code{bin_uri}) and \code{geography} fields to be available in the \code{bold_parquet_search} result. In case the \code{specific.cols} argument is used in the \code{bold_parquet_search} function to retrieve certain columns that do not have taxonomy and geography columns, the function will throw an error.
}
\examples{
\dontrun{


# Search the BOLD data package
bold_search <- bold_parquet_search(
  input.parquet = parquet_file,
  taxonomy = "Odonata",
  geography = "Thailand"
)

# Get the occurrence matrix

occurrence_data <- bcdm_to_occmatrix(
  bold_search,
  kingdom = "Animalia",
  taxon.rank = "family",
  site.cat = "region"
)
}
}
