% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/external.data.collect.R
\name{bold.data.collect}
\alias{bold.data.collect}
\title{Collect and export BOLD NODE search results}
\usage{
bold.data.collect(
  bold.search.res,
  chunk.size = 4e+06,
  sys.sleep = 0,
  export = FALSE,
  export.type = c("tsv", "parquet"),
  output.path = NULL
)
}
\arguments{
\item{bold.search.res}{A tbl_sql object obtained from 'bold.data.search'}

\item{chunk.size}{Maximum number of rows to process in each chunk (default: 4e6)}

\item{sys.sleep}{Time to sleep between chunks in seconds (default: 0)}

\item{export}{Logical value that allows user to export the output locally. Default value is FALSE}

\item{export.type}{Character string specifying the data type of the exported file (tsv or parquet)}

\item{output.path}{Character string specifying the local path for data export}
}
\value{
A data frame containing all collected results; if export = T, either a TSV or parquet file exported locally
}
\description{
collects,outputs and exports the tbl_sql 'bold.data.search' results object, processing large datasets in user defined manageable chunks.
}
\details{
This function processes large BOLD search results in manageable chunks to avoid memory issues.
It supports exporting results in either TSV or Parquet format. For Parquet export, data is written directly
without chunking using DuckDB's COPY command. For TSV export, data is collected in chunks and then written to  file.The function uses progress bars to track processing status and allows optional pausing between chunks via the
sys.sleep parameter. When export=FALSE, the function returns the collected data; when export=TRUE, it returns
NULL invisibly after writing the data to disk.#'
}
\examples{
\dontrun{


# Search the BOLD data
bold_search <- bold.data.search(
parquet_path=parquet_file,
taxonomy = "Coleoptera",
geography = "Canada",
marker = "COI-5P",
basecount = c(500, 660)
)

# Collect the data  (no export)
bold.data.collect(
bold_search,
chunk.size = 50000,
export = FALSE)

# Collect and export
bold.data.collect(
bold_search,
chunk.size = 50000,
export = TRUE,
export.type = "parquet",
output.path = 'userdefinedpath')

}
}
