% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/external.data.search.R
\name{bold.data.search}
\alias{bold.data.search}
\title{Search BOLD dataset release parquet data}
\usage{
bold.data.search(
  input.parquet,
  ids = NULL,
  bins = NULL,
  taxonomy = NULL,
  geography = NULL,
  institutes = NULL,
  identified.by = NULL,
  seq.source = NULL,
  marker = NULL,
  basecount = NULL,
  biogeo.cat = NULL,
  dataset.projects = NULL,
  bounding.box = NULL,
  ambi.base.cutoff = NULL,
  specific.cols = NULL
)
}
\arguments{
\item{input.parquet}{Path to the input parquet file}

\item{ids}{Vector of process IDs or sample IDs to filter by}

\item{bins}{Vector of BIN numbers (i.e. URIs) to filter by}

\item{taxonomy}{Vector of taxonomic names to filter by (can include kingdom, phylum, class, order, family, subfamily, genus, species)}

\item{geography}{Vector of geographic locations to filter by (can include country/ocean, province/state, region, sector, site)}

\item{institutes}{Vector of institute codes to filter by}

\item{identified.by}{Vector of identifiers to filter by}

\item{seq.source}{Vector of sequence run sites to filter by}

\item{marker}{Vector of marker codes to filter by}

\item{basecount}{Nucleotide base count filter - either a single value or a vector of two values for a range}

\item{biogeo.cat}{Vector of biogeographic/ecological categories to filter by (biome, realm, or ecoregion)}

\item{dataset.projects}{Vector of dataset/project codes to filter by}

\item{bounding.box}{Numeric vector of length 4: c(min_lon, max_lon, min_lat, max_lat)}

\item{ambi.base.cutoff}{Character value for filtering data based proportion of ambiguous bases (IUPAC codes).Three values currently available ("<1\%", "1-5\%", or ">5\%"); Default value is NULL}

\item{specific.cols}{Optional character vector of specific columns to return}
}
\value{
A tbl_sql object containing the filtered data
}
\description{
Search the BOLD dataset release based on various search criteria including taxonomy, geography, institutes etc.
}
\details{
This function loads a Parquet file containing BOLD dataset release data and applies filters based on the provided parameters.
It supports filtering by taxonomic identifiers (process IDs, sample IDs), taxonomy (from kingdom to species level),
geography (country/ocean to site level), institutes, identifiers, sequence sources, genetic markers, nucleotide base counts,
biogeographic categories, dataset/projects, bounding boxes, and ambiguous base cutoffs. The function returns a tbl_sql object
that represents the filtered data, allowing for efficient querying without loading the entire dataset into memory.
Users can also specify particular columns to return using the specific.cols parameter.
}
\examples{
\dontrun{


# Search the BOLD data

# Taxonomy
bold_search <- bold.data.search(
parquet_path=parquet_file,
taxonomy = c("Odonata","Poecilia")
)

# Geography
bold_search <- bold.data.search(
parquet_path=parquet_file,
geography = "Canada"
)

# Combination of many search criteria
bold_search <- bold.data.search(
parquet_path=parquet_file,
taxonomy = "Coleoptera",
geography = "Canada",
marker = "COI-5P",
basecount = c(500, 660)
)


}

}
