% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/external.get.vocab.R
\name{bold.get.vocab}
\alias{bold.get.vocab}
\title{Extract unique vocabulary from a BOLD dataset}
\usage{
bold.get.vocab(
  input.data,
  specific.cols,
  save.data = FALSE,
  output.file = NULL
)
}
\arguments{
\item{input.data}{Path to the input parquet file or the \code{bold.data.search} result}

\item{specific.cols}{Name of the column to extract unique values from}

\item{save.data}{Logical indicating whether to save the results to disk as a .rds file (default: FALSE)}

\item{output.file}{Path (without extension) for saving results as .rds file (required if save.data = TRUE)}
}
\value{
A list containing unique values from the specified column; if \code{save.data} = T, a \code{.rds} file exported locally
}
\description{
Extracts distinct values of specified field/s in BOLD parquet data
}
\details{
This function extracts unique values from one or more specified columns in BOLD parquet data. It handles both the parquet file and \code{tbl_sql} objects from \code{bold.data.search} as input. The results can be saved to disk as an RDS file for later use.
}
\examples{
\dontrun{


# Search the BOLD data
bold_search <- bold.data.search(
input.parquet=parquet_file,
taxonomy = "Diptera",
geography = "India",
marker = "COI-5P")

# Get the vocabulary

vocab.data <- bold.get.vocab(bold_search,bold_search,specific.cols = c("inst","identified.by"))

}
}
