% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/external.get.DwC.R
\name{get.dwc}
\alias{get.dwc}
\title{Convert bold.data.search results in to a Darwin Core (DwC) format data model}
\usage{
get.dwc(bold.search.res)
}
\arguments{
\item{bold.search.res}{A \code{tbl_sql} object containing BOLD search results}
}
\value{
A data frame with columns mapped to Darwin Core equivalent fields
}
\description{
Converts \code{bold.data.search} results from BCDM format to Darwin Core Standard format.
}
\details{
This function maps BCDM (Barcode Core Data Model) fields to their Darwin Core equivalents.
\emph{Important Note}: All fields should be available in the \code{bold.data.search} \code{tbl_sql} object otherwise, the function will throw an error
}
\examples{
\dontrun{

# Search the BOLD data
bold_search <- bold.data.search(
input.parquet=parquet_file,
taxonomy = c("Cerambycidae","Meloidae"),
geography = "Australia"
)

# Get the DNAStringset object

bold.dwc<-get.dwc(bold_search)

}
}
