% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/external.get.fasta.R
\name{get.fasta}
\alias{get.fasta}
\title{Export BOLD search results to FASTA format}
\usage{
get.fasta(bold.search.res, output.file, fas.header, chunk.size = 1e+06)
}
\arguments{
\item{bold.search.res}{A tbl_sql object containing BOLD search results}

\item{output.file}{Path to the output FASTA file}

\item{fas.header}{Character vector of field names to include in the FASTA header (separated by "|")}

\item{chunk.size}{Number of records to process in each chunk (default: 1e6)}
}
\value{
writes a FASTA file to disk with custom headers as specified by the user
}
\description{
Exports nucleotide sequences from BOLD search results to a FASTA file with customizable headers.
}
\details{
This function exports nucleotide sequences from BOLD search results to a FASTA file, which is a standard format for sequence data. It processes the data in chunks to handle large datasets efficiently, constructs FASTA headers from user-specified fields (joined with "|"), and writes the data directly to disk. The function uses DuckDB's SQL capabilities to safely concatenate fields for headers and employs temporary tables for repeated operations. Progress is tracked through chunk-based processing.
}
\examples{
\dontrun{


# Search the BOLD data
bold_search <- bold.data.search(
parquet_path=parquet_file,
taxonomy = "Coleoptera",
geography = "Canada",
marker = "COI-5P",
basecount = c(500, 660)
)

# Get a fasta file
get.fasta(
bold_search,
output.file = "trial2.fas",
fas.header = c("bin_uri", "processid")
)


}

}
