% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/external.get.sf.R
\name{get.sf}
\alias{get.sf}
\title{Convert BOLD search results to spatial points}
\usage{
get.sf(bold.search.res, chunk.size = 10000)
}
\arguments{
\item{bold.search.res}{A tbl_sql object containing BOLD search results}

\item{chunk.size}{Number of records to process in each chunk (default: 10000)}
}
\value{
An sf object with point geometry in WGS84 coordinate system (EPSG:4326)
}
\description{
Converts BOLD search results with coordinate data to spatial points (sf object).
}
\details{
This function converts BOLD search results containing coordinate data into spatial points using the Simple Features (sf) package. It processes coordinate strings by extracting and cleaning the values, then creates point geometries in the WGS84 coordinate system (EPSG:4326). The function uses chunking to handle large datasets efficiently, processing coordinates in batches to avoid memory issues. Invalid or missing coordinates are filtered out during processing.
}
\examples{
\dontrun{


# Search the BOLD data
bold_search <- bold.data.search(
parquet_path=parquet_file,
taxonomy = "Odonata",
geography = "Malaysia")

# Get the occurrence matrix

occurrence_data <- get.sf(bold_search, chunk.size = 10000)

}

}
