% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/external.get_concise_summary.R
\name{get_concise_summary}
\alias{get_concise_summary}
\title{Generate a concise summary of the search results}
\usage{
get_concise_summary(bold.search.res)
}
\arguments{
\item{bold.search.res}{A \code{tbl_sql} object containing \code{bold_parquet_search} results.}
}
\value{
A data frame with the summary statistics.
}
\description{
Creates a summary statistics table from the \code{bold_parquet_search} \code{tb_sql} object.
}
\details{
The function provides a concise summary of the search obtained by \code{bold_parquet_search} that includes: total records, unique BINs, unique institutes, unique markers and amplicon size range.
}
\examples{
\dontrun{

# Search the BOLD data package

parquet_file <- "user defined path to parquet file"

bold_search <- bold_parquet_search(
  input.parquet = parquet_file,
  taxonomy = "Hemiptera",
  geography = "India",
  marker = "COI-5P"
)
# Get the concise summary
bold_summary <- get_concise_summary(bold_search)
}
}
