% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/external.data.collect.R
\name{bold.data.collect}
\alias{bold.data.collect}
\title{Collect and export BOLD NODE search results}
\usage{
bold.data.collect(
  bold.search.res,
  chunk.size = 1e+06,
  sys.sleep = 0,
  export = FALSE,
  export.type = c("tsv", "parquet"),
  output.path = NULL
)
}
\arguments{
\item{bold.search.res}{A \code{tbl_sql} object obtained from 'bold.data.search'}

\item{chunk.size}{Maximum number of rows to process in each chunk. Default value is 1e6}

\item{sys.sleep}{Time to sleep between chunks in seconds. Default value is 0}

\item{export}{Logical value that allows user to export the output locally. Default value is FALSE}

\item{export.type}{Character string specifying the data type of the exported file (tsv or parquet)}

\item{output.path}{Character string specifying the local path for data export along with the file name and extension}
}
\value{
A data frame containing all collected results; if export = T, either a TSV or parquet file exported locally
}
\description{
collects,outputs and exports the tbl_sql 'bold.data.search' results object, processing large datasets in user defined manageable chunks.
}
\details{
This function collects (downloads it in the R session) the search results from \code{bold.data.search}. Data chunking and system sleep options are available to manage large sizes to avoid memory issues. The function also supports exporting the searches in either TSV or Parquet format. export=FALSE (default) returns the collected data. The full output_path has to be provided (along with the intended file name and file extension) when export=TRUE.
\emph{Important Note}: Some data searches (e.g. all Diptera) once collected can get very large and overload the RAM capacity for some machines.
}
\examples{
\dontrun{


# Search the BOLD data
bold_search <- bold.data.search(
input.parquet=parquet_file,
taxonomy = "Coleoptera",
geography = "Canada",
marker = "COI-5P",
basecount = c(500, 660)
)

# Collect the data  (no export)
bold.data.collect(
bold_search,
chunk.size = 50000,
export = FALSE)

# Collect and export
bold.data.collect(
bold_search,
chunk.size = 50000,
export = TRUE,
export.type = "parquet",
output.path = 'userdefinedpath')

}
}
