% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/external.get.vocab.R
\name{bold.get.vocab}
\alias{bold.get.vocab}
\title{Extract unique vocabulary from a BOLD dataset}
\usage{
bold.get.vocab(
  input.data,
  specific.cols,
  save.data = FALSE,
  output.file = NULL
)
}
\arguments{
\item{input.data}{Path to the input parquet file or the \code{bold.data.search} result}

\item{specific.cols}{Name of the column to extract unique values from}

\item{save.data}{Logical indicating whether to save the results to disk as a .rds file (default: FALSE)}

\item{output.file}{Path (without extension) for saving results as .rds file (required if save.data = TRUE)}
}
\value{
A list containing unique values from the specified column; if \code{save.data} = T, a rds file exported locally
}
\description{
Extracts distinct values of specified field/s in BOLD parquet data
}
\details{
This function extracts unique values from one or more specified columns in BOLD parquet data. It handles both direct file paths and tbl_sql objects as input, making it flexible for use in different workflows. The function filters out missing and empty values before extracting distinct values, ensuring clean vocabulary lists. Optionally, the results can be saved to disk as an RDS file for later use. This is particularly useful for enumerating possible values in categorical fields like taxon names, geographic locations, or genetic markers.
}
\examples{
\dontrun{


# Search the BOLD data
bold_search <- bold.data.search(
parquet_path=parquet_file,
taxonomy = "Diptera",
geography = "India",
marker = "COI-5P")

# Get the vocabulary

vocab.data <- bold.get.vocab(bold_search,bold_search,specific.cols = c("inst","identified.by"))

}
}
