% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/external.bold_parquet_search.R
\name{bold_parquet_search}
\alias{bold_parquet_search}
\title{Search records within a BOLD parquet data package}
\usage{
bold_parquet_search(
  input.parquet,
  ids = NULL,
  bins = NULL,
  scope.taxonomy = NULL,
  taxonomy = NULL,
  scope.geography = "any",
  geography = NULL,
  institutes = NULL,
  identified.by = NULL,
  seq.source = NULL,
  marker = NULL,
  basecount = NULL,
  biogeo.cat = NULL,
  dataset.projects = NULL,
  bounding.box = NULL,
  ambi.base.cutoff = NULL,
  specific.cols = NULL
)
}
\arguments{
\item{input.parquet}{Path to the input parquet file.}

\item{ids}{Vector of process IDs or sample IDs used to filter records.}

\item{bins}{Vector of BIN numbers (i.e., URIs) used to filter records.}

\item{scope.taxonomy}{Character value specifying the kingdom (includes "all", "Animalia", "Plantae", "Protista", "Fungi", "Bacteria") (default: NULL).}

\item{taxonomy}{Vector of taxonomic names to filter by (includes "kingdom", "phylum", "class", "order", "family", "subfamily","genus", "species").}

\item{scope.geography}{A single character value specifying the geographic hierarchy level to search within (includes "any", "country.ocean", "province.state", "region", "sector", "site"; default: "any")}

\item{geography}{Vector of geographic locations to filter by based on the \code{scope.geography}.}

\item{institutes}{Vector of institute codes used to filter records.}

\item{identified.by}{Vector of identifiers used to filter records.}

\item{seq.source}{Vector of sequence run sites used to filter records.}

\item{marker}{Vector of marker codes used to filter records.}

\item{basecount}{Nucleotide base count filter - either a single value or a vector of two values for a range.}

\item{biogeo.cat}{Vector of biogeographic/ecological categories to filter by (biome, realm, or ecoregion).}

\item{dataset.projects}{Vector of dataset/project codes used to filter records.}

\item{bounding.box}{Numeric vector of length 4: c(min_lon, max_lon, min_lat, max_lat).}

\item{ambi.base.cutoff}{Character value for filtering data based proportion of ambiguous bases (IUPAC codes). Three values currently available ("<1\%", "1-5\%", or ">5\%") (default: NULL).}

\item{specific.cols}{Optional character vector of specific columns to return.}
}
\value{
A \code{tbl_sql} object containing the filtered data. Total records in the search are printed on the console.
}
\description{
Query records in BOLD parquet data packages using taxonomic, geographic, and other search criteria.
}
\details{
This function loads the BOLD public data packages parquet files (\url{https://boldsystems.org/data/data-packages/}) via DuckDB and applies filters based on the provided parameters. It supports filtering by ids (sampleid, processid), taxonomy (using combination of scope.taxonomy and taxonomy), geography (using combination of scope.geography and geography), biogeography (biome to ecoregion level), BINs, institutes, identifiers, sequence sources, genetic markers, nucleotide base counts, dataset or projects, spatial bounding boxes, and ambiguous base percent cutoffs. The taxonomy and geography filters use a two-level scoping system that allows users to define \code{scope.taxonomy} and \code{scope.geography} to control how records are filtered. This is particularly useful when the same name exists at multiple geographic or taxonomic levels. For example, the name Azerbaijan may refer to a country, but a region with the same name also exists in Iran. If \code{scope.geography} is set to \code{any}, records from both geographic levels will be returned. However, setting \code{scope.geography} to \code{country.ocean} restricts the search to country records only, returning records associated with the country of Azerbaijan. A similar situation can occur with taxonomic names where identical names are assigned to different groups. For example, the genus Iris occurs in both plants and animals. Using \code{scope.taxonomy} allows users to specify the desired taxonomic context and avoid ambiguity between groups. Users can also specify particular columns to return using the \code{specific.cols} parameter (column names can be checked using the \code{bcdm_field_names} function). The \code{tbl_sql} object can then be used by any of the \verb{bcdm_to_*} functions for data transformations or \code{bold_search_collect} to load all the data in memory.
}
\examples{
\dontrun{


# Search the BOLD data package

# Taxonomy
bold_search <- bold_parquet_search(
  input.parquet = parquet_file,
  taxonomy = c("Odonata", "Poecilia")
)

# Geography
bold_search <- bold_parquet_search(
  input.parquet = parquet_file,
  geography = "Canada"
)

# Combination of many search criteria
bold_search <- bold_parquet_search(
  input.parquet = parquet_file,
  scope.taxonomy = 'Animalia',
  taxonomy = "Coleoptera",
  scope.geography = 'country.ocean'
  geography = "Canada",
  marker = "COI-5P",
  basecount = c(500, 660)
)
}

}
