% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/external.get.concise.summary.R
\name{get.concise.summary}
\alias{get.concise.summary}
\title{Generate a concise summary of bold.data.search results}
\usage{
get.concise.summary(bold.search.res)
}
\arguments{
\item{bold.search.res}{A \code{tbl_sql} object containing \code{bold.data.search} results}
}
\value{
A data frame with the summary statistics
}
\description{
Creates a summary statistics table from the \code{bold.data.search} \code{tb_sql} object.
}
\details{
The function provides a concise summary of the search obtained by \code{bold.data.search} that includes details like total records, unique BINs, unique institutes, unique markers and amplicon size range.
}
\examples{
\dontrun{

# Search data

parquet_file<-'user defined path to parquet file'

bold_search <- bold.data.search(
input.parquet=parquet_file,
taxonomy = "Hemiptera",
geography = "India",
marker = "COI-5P"
)
# Get the concise summary
bold_summary <- get.concise.summary(bold_search)

}
}
