% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/external.get.DNAstringset.R
\name{get.DNAStringSet}
\alias{get.DNAStringSet}
\title{Convert the BOLD search to a DNAStringSet object}
\usage{
get.DNAStringSet(bold.search.res, marker = NULL, cols_for_seq_names)
}
\arguments{
\item{bold.search.res}{A tbl_sql object containing BOLD search results}

\item{marker}{Character vector specifying the genetic marker}

\item{cols_for_seq_names}{Character vector of field names to include in the header (separated by "|")}
}
\value{
A DNAStringSet object
}
\description{
Converts the sequence data from the BOLD search into a DNAStringSet object for downstream multiple sequence alignment with customized headers.
}
\details{
This function processes BOLD search results to extract nucleotide sequences, clean them by removing gap characters, and optionally filter by specific genetic markers. It constructs sequence names from user-specified fields and returns a named DNAStringSet object suitable for multiple sequence alignment or other phylogenetic analyses. The function works directly with tbl_sql objects to handle large datasets efficiently.
}
\examples{
\dontrun{


# Search the BOLD data
bold_search <- bold.data.search(
parquet_path=parquet_file,
taxonomy = "Coleoptera",
geography = "Canada",
basecount = c(500, 660)
)

# Get the DNAStringset object

bold.dnastringset<-get.DNAStringSet(bold_search,
marker="COI-5P",
cols_for_seq_names = c("processid","family")
)

}

}
