% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/external.get.occ.data.R
\name{get.occ.data}
\alias{get.occ.data}
\title{Extract occurrence data from BOLD search results}
\usage{
get.occ.data(
  bold.search.res,
  kingdom = "Animalia",
  taxon.rank,
  taxon.name = NULL,
  site.cat = NULL,
  pre.abs = FALSE
)
}
\arguments{
\item{bold.search.res}{A \code{tbl_sql} object containing BOLD search results}

\item{taxon.rank}{Taxonomic rank to aggregate by (kingdom, phylum, class, order, family, genus, or species)}

\item{taxon.name}{Optional vector of specific taxon names to include}

\item{site.cat}{Optional categorical variable to group occurrence data by (e.g., region)}

\item{pre.abs}{Logical indicating whether to convert counts to presence/absence (1/0) data (default: FALSE)}
}
\value{
A data frame with occurrence data (taxon names as columns, site categories or coordinates as rows)
}
\description{
Extracts occurrence data (specimen counts by taxon and location) from BOLD search results.
}
\details{
This function transforms the search results from \code{bold.data.search} into occurrence data matrices used commonly in biodiversity and ecological analyses by packages like \code{vegan} and \code{betapart}. \code{kingdom} argument specifies the taxa kingdom with the default value being \code{Animalia}. The occurrences differ based on the kingdom. For \code{Animalia}, only records having BINs are counted (i.e. records without BINs are removed before calculations), while for other kingdoms, all records having a sequence are counted (i.e. records without sequences are removed before calculations). It supports aggregation at different taxonomic ranks (from kingdom to species) for a single or multiple taxa (also applicable for bin_uri; the difference being that the numbers in each cell would be the number of times that respective BIN is found at a particular \code{site.cat}), optional filtering by specific taxon names. \code{site.cat} can be any of the \code{geography} fields (Meta data on fields can be checked using the \code{bold.bcdm.fields} function). In case where \code{site.cat} = NULL, the column \code{coord} will be used as default. The function can convert count data to presence/absence (1/0) format. In case \code{specific.cols} is used in the \code{bold.data.search} function to obtain only few columns, occurrence matrix wont be generated.
}
\examples{
\dontrun{


# Search the BOLD data
bold_search <- bold.data.search(
input.parquet=parquet_file,
taxonomy = "Odonata",
geography = "Thailand")

# Get the occurrence matrix

occurrence_data <- get.occ.data(
  bold_search,
  kingdom = 'Animalia',
  taxon.rank = "family",
  site.cat = "region"
  )

}
}
