% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/external.get.sf.R
\name{get.sf}
\alias{get.sf}
\title{Convert BOLD search results to spatial points}
\usage{
get.sf(bold.search.res, chunk.size = 1e+05)
}
\arguments{
\item{bold.search.res}{A tbl_sql object containing BOLD search results}

\item{chunk.size}{Number of records to process in each chunk (default: 10000)}
}
\value{
An sf object with point geometry in WGS84 coordinate system (EPSG:4326)
}
\description{
Converts BOLD search results with coordinate data to spatial points (sf object).
}
\details{
This function transforms the search results from \code{bold.data.search} into an \code{sf} object. Data chunking  option is available to manage large sizes to avoid memory issues. The function creates point geometries in the WGS84 coordinate system (EPSG:4326). Records that dont have coordinate data are removed during processing.
}
\examples{
\dontrun{


# Search the BOLD data
bold_search <- bold.data.search(
input.parquet=parquet_file,
taxonomy = "Odonata",
geography = "Malaysia")

# Get the occurrence matrix

occurrence_data <- get.sf(bold_search, chunk.size = 100000)

}

}
