% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/external.get_bin_reps.R
\name{get_bin_reps}
\alias{get_bin_reps}
\title{Select representative records by BIN}
\usage{
get_bin_reps(
  bold.search.res,
  Nreps = 1,
  by.taxon = FALSE,
  non.redundant.taxa = FALSE,
  enforce.scientific = FALSE,
  criteria = list(vouchered = TRUE,
                  seq_length = c("COI_auto", "longest", "shortest", 658),
                  id_method = c("Morphology", "Morphology and sequence based",
                                "Image based", "Image and sequence based",
                                "Tree based", "BIN based", "BOLD ID Engine",
                                "Other sequence based approach", "Other"),
                 inst = "Centre for Biodiversity Genomics",
                 coll_date = c("latest", "oldest"),
                 seq_date = c("latest", "oldest")),
  seed = NULL
)
}
\arguments{
\item{bold.search.res}{A \code{tbl_sql} object obtained from \code{\link{bold_parquet_search}} or a data frame or data table in BCDM format.}

\item{Nreps}{Integer indicating the maximum number of representatives to select for each BIN (or BIN-taxon combination).}

\item{by.taxon}{Logical value indicating whether to select representatives for each unique combination of BIN and taxonomic identification. If \code{TRUE}, the additional parameters \code{non_redundant_taxa} and \code{enforce_scientific} are also applied.}

\item{non.redundant.taxa}{Logical value indicating whether to select representatives at the lowest available rank from each distinct taxonomic lineage. For example, in a BIN with the identifications "Apidae" and "Bombus impatiens", only records assigned to "Bombus impatiens" will be selected. Ignored if \code{by.taxon} is \code{FALSE}.}

\item{enforce.scientific}{Logical value indicating whether to ignore representatives with non-scientific, provisional names. Ignored if \code{by.taxon} is \code{FALSE}.}

\item{criteria}{Named list of selection criteria to apply when sampling representatives, given in priority order. See details for more information and default values.}

\item{seed}{Optional positive integer to use as a random seed for reproducible tie-breaking. If \code{NULL} (the default), ties are broken randomly and selected records may differ between runs.}
}
\value{
A data frame of selected representatives.
}
\description{
Obtain one or more representative record(s) from each BIN in search results or
a BCDM data frame. BIN representatives can also be selected for each unique BIN-taxon
combination. The provided selection criteria are applied in sequence to select
representatives, thus criteria should be given in order of priority. If multiple
records are tied after all criteria have been applied, the tie can be broken
randomly (the default) or deterministically by setting a random seed, making it
possible to obtain reproducible results for the same input values.
}
\details{
The provided \code{bold.search.res} input can be a search result object from
\code{\link{bold_parquet_search}} or a BCDM data frame. Alternatively, it can be
any data frame or data table minimally containing \code{bin_uri}, unique record identifiers
(e.g. \code{processid} or \code{sampleid}), taxonomic identifications for all available records,
and any fields relevant to the provided selection \code{criteria} (see below).

Selection \code{criteria} must be listed in order of priority. Available criteria
include the following:
\itemize{
\item \code{vouchered}: If \code{TRUE}, prioritize records with known voucher
repositories over those mined from databases like GenBank. Setting this to
\code{FALSE} will prioritize records \emph{without} vouchers. To exclude this
criterion, simply omit it. Corresponding BCDM field: \code{inst}.
\item \code{seq_length}: Can be used either to specify target barcode sequence
length as an integer, to preferentially select longer or shorter sequences
("longest", "shortest"), or to select sequences that match the modal* barcode
length for each BIN ("COI_auto"). If a target length is provided, sequences
closest to that target are prioritized. *Modal barcode length ("COI_auto"
option) is determined after first rounding sequence lengths to the nearest
full codon; in the event of multiple modes, the one closest to 658bp is
chosen. Corresponding BCDM field: \code{nuc_basecount}.
\item \code{id_method}: A character vector listing preferred identification methods
in order of priority. The function definition lists all available values per
the BCDM specification. Corresponding BCDM field: \code{identification_method}.
\item \code{inst}: A character vector listing preferred voucher specimen repositories
in order of priority. Values not present in the input data are ignored.
Corresponding BCDM field: \code{inst}.
\item \code{coll_date}: Prioritize recently collected specimens ("latest") or
those collected longest ago ("oldest"). Records without dates are selected
last in either case. Corresponding BCDM field: \code{collection_date_start}.
\item \code{seq_date}: Prioritize recently uploaded sequences ("latest") or
those with the earliest upload date ("oldest"). Corresponding BCDM field:
\code{sequence_upload_date}.
}

Default selection criteria are as follows:

\if{html}{\out{<div class="sourceCode R">}}\preformatted{criteria = list(vouchered = TRUE,
                seq_length = "COI_auto",
                id_method = c("Morphology", "Morphology and sequence based",
                              "Image based", "Image and sequence based",
                              "Tree based", "BIN based", "BOLD ID Engine",
                              "Other sequence based approach", "Other"),
                inst = "Centre for Biodiversity Genomics",
                coll_date = "latest",
                seq_date = "latest")
}\if{html}{\out{</div>}}

\strong{Note}: The same data provided either as a \code{tbl_sql} object or a data frame
may yield different representative records, even with the same random seed.
This is because input goes through one of two paths depending on format, each
with different default tie-breaking behaviours.
}
\examples{
\dontrun{

# Search BOLD data package
bold_search <- bold_parquet_search(
  input.parquet = parquet_file,
  taxonomy = "Araneae",
  geography = "Canada"
)

# Select three representatives per BIN from the searched data,
# prioritizing those with a morphological ID and with vouchers
# deposited at the CBG (e.g. in case vouchers need to be examined)
bin_reps <- get_bin_reps(
  bold.search.res = bold_search,
  Nreps = 3,
  criteria = list(inst = "Centre for Biodiversity Genomics",
                  id_method = c("Morphology",
                                "Morphology and sequence based"))
)

# Select one representative for each combination of BIN and taxonomic 
# lineage (scientific names only), with preference for 658-bp barcodes
# (e.g. for building a sequence tree of all known taxa)
bin_tax_reps <- get_bin_reps(
  bold.search.res = bold_search,
  Nreps = 1,
  by.taxon = TRUE,
  non.redundant.taxa = TRUE,
  enforce.scientific = TRUE,
  criteria = list(seq_length = 658)
)
}

}
